% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_sample_means.R
\name{many_sample_means}
\alias{many_sample_means}
\title{perform the sample_mean process many times
and return a vector of the results.}
\usage{
many_sample_means(vec, n, reps)
}
\arguments{
\item{vec}{A vector}

\item{n}{An integer}

\item{reps}{An integer}
}
\value{
a vector of the result of calling sample_mean many times
}
\description{
perform the sample_mean process many times
and return a vector of the results.
}
