% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{sample_mean}
\alias{sample_mean}
\title{take as input a vector vec and an integer n.
It should take a random sample of size n from vec,
then calculate and return the mean of that subsample.}
\usage{
sample_mean(vec, n)
}
\arguments{
\item{vec}{A vector}

\item{n}{An integer}
}
\value{
an int, mean of subsample
}
\description{
take as input a vector vec and an integer n.
It should take a random sample of size n from vec,
then calculate and return the mean of that subsample.
}
