% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_means_ns.R
\name{sample_means_ns}
\alias{sample_means_ns}
\title{Perform the many_sample_means process
for each of the values in the ns vector.}
\usage{
sample_means_ns(vec, reps, ns)
}
\arguments{
\item{vec}{A vector}

\item{reps}{An integer}

\item{ns}{A vector of ns}
}
\value{
a dataframe of calling many_sample_means
}
\description{
Perform the many_sample_means process
for each of the values in the ns vector.
}
